+++ legwork/tests/test_sources.py
@@ -27,7 +27,7 @@ class Test(unittest.TestCase):
                                 interpolate_sc=False)
 
         # calculate the merger times
-        t_merge = sources.get_merger_time()
+        t_merge = sources.get_merger_time(save_in_class=False)
 
         # create a new class after evolving every source for 10 years
         evolved_sources = sources.evolve_sources(10 * u.yr,
@@ -238,7 +238,7 @@ class Test(unittest.TestCase):
         m_1 = np.random.uniform(0, 10, n_values) * u.Msun
         m_2 = np.random.uniform(0, 10, n_values) * u.Msun
         dist = np.random.uniform(0, 30, n_values) * u.kpc
-        f_orb = 10**(np.random.uniform(-5, -1, n_values)) * u.Hz
+        f_orb = 10**(np.random.uniform(-5, -3, n_values)) * u.Hz
         ecc = np.random.uniform(0.0, 0.9, n_values)
 
         # compare snr calculated directly with through Source
@@ -283,7 +283,7 @@ class Test(unittest.TestCase):
         n_values = 10
         m_1 = np.random.uniform(0, 10, n_values) * u.Msun
         m_2 = np.random.uniform(0, 10, n_values) * u.Msun
-        ecc = np.random.uniform(0.0, 1.0, n_values)
+        ecc = np.random.uniform(0.0, 0.95, n_values)
         dist = np.random.uniform(0, 10, n_values) * u.kpc
         f_orb = 10**(np.random.uniform(-5, -1, n_values)) * u.Hz